fix: properly set frame app context on reload
