+++ .changeset/popular-bats-relate.md
@@ -1,0 +1,5 @@
+---
+"@frames.js/debugger": patch
+---
+
+fix: properly set frame app context on reload
